+++ turbo/apps/platform/src/views/okou-page/chat-thread-page.tsx
@@ -97,12 +97,7 @@ import type {
   UserMessageDocument,
   UserMessagePart,
 } from "@okouai/api-contracts/contracts/chat-threads";
-import {
-  chatEventCompatibilityRole,
-  foldLatestChatUsageByRunId,
-  isChatEventContentTextType,
-  terminatedChatRunIds,
-} from "@okouai/api-contracts/contracts/chat-events";
+import { isChatEventContentTextType } from "@okouai/api-contracts/contracts/chat-events";
 import {
   messageDocumentToDisplayText,
   messageDocumentToPrompt,
@@ -152,10 +147,16 @@ import {
   closeChatConnectorActionConnectDialog$,
 } from "../../signals/chat-page/connector-action-block.ts";
 import {
+  buildCompletedWorkFolding,
+  chatEventDisplayError,
   completedWorkExpandedKeys$,
+  completedWorkExpandedKeysForScrollTarget,
+  completedWorkFoldForGroup,
+  isRenderableAssistantEvent,
   toggleCompletedWorkExpanded$,
+  type CompletedWorkFold,
+  type CompletedWorkFolding,
 } from "../../signals/chat-page/completed-work-folding.ts";
-import { isCancelledRunEvent } from "../../signals/chat-page/chat-run-lifecycle.ts";
 import {
   buildRunGroupFolding,
   runGroupExpansionOverrides$,
@@ -489,24 +490,6 @@ function eventNonContentPart(
   return userMessageNonContentPart(
     isInputChatEvent(event) ? event.userMessage : undefined,
   );
-}
-
-function chatEventAttachments(event: ChatEvent) {
-  return isInputChatEvent(event)
-    ? userMessageFileAttachments(event.userMessage)
-    : undefined;
-}
-
-function chatEventError(event: ChatEvent): string | undefined {
-  if (
-    event.eventType === "input.rejected" ||
-    event.eventType === "output.error" ||
-    event.eventType === "run.failed" ||
-    event.eventType === "run.cancelled"
-  ) {
-    return event.error;
-  }
-  return undefined;
 }
 
 function ArtifactsButton({ thread }: { thread: ChatPanelSignals }) {
@@ -3482,367 +3465,12 @@ function assistantGroupIdForCollapsedRunGroupFold(
   return undefined;
 }
 
-function completedWorkFoldForGroup(
-  completedWorkFolding: CompletedWorkFolding | null,
-  group: ChatEventGroup,
-): CompletedWorkFold | null {
-  if (completedWorkFolding === null) {
-    return null;
-  }
-  return (
-    group.events
-      .map((event) => {
-        return completedWorkFolding.foldsByFinalEventId.get(event.id);
-      })
-      .find((fold) => {
-        return fold !== undefined;
-      }) ?? null
-  );
-}
-
-function groupEventsByRole(
-  events: readonly EnrichedChatEvent[],
-): ChatEventGroup[] {
-  const groups: ChatEventGroup[] = [];
-  for (const event of events) {
-    const role = chatEventCompatibilityRole(event.eventType);
-    const last = groups[groups.length - 1];
-    if (last && last.role === role) {
-      last.events.push(event);
-      continue;
-    }
-    groups.push({
-      beginEventId: event.id,
-      role,
-      events: [event],
-    });
-  }
-  return groups;
-}
-
-interface CompletedWorkFold {
-  key: string;
-  finalEventId: string;
-  hiddenGroups: ChatEventGroup[];
-  labelGroups: ChatEventGroup[];
-}
-
-interface CompletedWorkFolding {
-  visibleGroups: ChatEventGroup[];
-  foldsByFinalEventId: Map<string, CompletedWorkFold>;
-}
-
-function completedWorkExpandedKeysForScrollTarget(
-  folding: CompletedWorkFolding | null,
-  expandedKeys: ReadonlySet<string>,
-  targetEventId: string | null,
-): ReadonlySet<string> {
-  if (folding === null || targetEventId === null) {
-    return expandedKeys;
-  }
-  const targetFold = Array.from(folding.foldsByFinalEventId.values()).find(
-    (fold) => {
-      return fold.hiddenGroups.some((group) => {
-        return group.events.some((event) => {
-          return event.id === targetEventId;
-        });
-      });
-    },
-  );
-  if (!targetFold || expandedKeys.has(targetFold.key)) {
-    return expandedKeys;
-  }
-  const next = new Set(expandedKeys);
-  next.add(targetFold.key);
-  return next;
-}
-
-function groupEventsForCompletedWorkDisplay(
-  events: readonly EnrichedChatEvent[],
-  foldFinalEventIds: ReadonlySet<string>,
-): ChatEventGroup[] {
-  const groups: ChatEventGroup[] = [];
-  for (const event of events) {
-    const role = chatEventCompatibilityRole(event.eventType);
-    const forceStandalone = foldFinalEventIds.has(event.id);
-    const last = groups[groups.length - 1];
-    const lastHasFoldFinal =
-      last?.events.some((candidate) => {
-        return foldFinalEventIds.has(candidate.id);
-      }) ?? false;
-    const lastFoldFinal = last?.events.find((candidate) => {
-      return foldFinalEventIds.has(candidate.id);
-    });
-    const continuesFoldFinalRun =
-      lastFoldFinal?.runId !== undefined && lastFoldFinal.runId === event.runId;
-
-    if (
-      !forceStandalone &&
-      last &&
-      last.role === role &&
-      (!lastHasFoldFinal || continuesFoldFinalRun)
-    ) {
-      last.events.push(event);
-      continue;
-    }
-
-    groups.push({
-      beginEventId: event.id,
-      role,
-      events: [event],
-    });
-  }
-  return groups;
-}
-
 function firstRunIdForEvents(
   events: readonly EnrichedChatEvent[],
 ): string | undefined {
   return events.find((event) => {
     return event.runId !== undefined;
   })?.runId;
-}
-
-function usageByRunIdFromGroups(
-  groups: readonly ChatEventGroup[],
-): Map<string, ChatEventUsagePayload> {
-  return foldLatestChatUsageByRunId(
-    groups.flatMap((group) => {
-      const runId = firstRunIdForEvents(group.events);
-      return group.role === "assistant" &&
-        group.usage !== undefined &&
-        runId !== undefined
-        ? [
-            {
-              eventType: "usage.recorded" as const,
-              runId,
-              usage: group.usage,
-            },
-          ]
-        : [];
-    }),
-  );
-}
-
-function attachUsageToCompletedWorkGroups(
-  groups: readonly ChatEventGroup[],
-  usageByRunId: ReadonlyMap<string, ChatEventUsagePayload>,
-): ChatEventGroup[] {
-  const lastAssistantGroupIndexByRunId = new Map<string, number>();
-  for (const [index, group] of groups.entries()) {
-    if (
-      group.role !== "assistant" ||
-      !group.events.some(isRenderableAssistantEvent)
-    ) {
-      continue;
-    }
-    const runId = firstRunIdForEvents(group.events);
-    if (runId !== undefined) {
-      lastAssistantGroupIndexByRunId.set(runId, index);
-    }
-  }
-  return groups.map((group, index) => {
-    if (group.role !== "assistant") {
-      return group;
-    }
-    const runId = firstRunIdForEvents(group.events);
-    if (
-      runId === undefined ||
-      lastAssistantGroupIndexByRunId.get(runId) !== index
-    ) {
-      return group;
-    }
-    const usage = usageByRunId.get(runId);
-    return usage === undefined ? group : { ...group, usage };
-  });
-}
-
-function isRenderableAssistantEvent(event: EnrichedChatEvent): boolean {
-  return (
-    chatEventCompatibilityRole(event.eventType) === "assistant" &&
-    ((isChatEventContentTextType(event.eventType) && Boolean(event.content)) ||
-      Boolean(chatEventError(event)) ||
-      hasChatEventBodyContent(event) ||
-      Boolean(chatEventAttachments(event)?.length))
-  );
-}
-
-function isThinkingOnlyAssistantEvent(event: EnrichedChatEvent): boolean {
-  return (
-    event.eventType === "output.thinking" && event.thinking.trim().length > 0
-  );
-}
-
-function terminatedRunIdsForCompletedWork(
-  events: readonly EnrichedChatEvent[],
-): Set<string> {
-  return terminatedChatRunIds(events);
-}
-
-function splitCompletedWorkEventsAtUsers(
-  events: readonly EnrichedChatEvent[],
-): EnrichedChatEvent[][] {
-  const phases: EnrichedChatEvent[][] = [];
-  let phase: EnrichedChatEvent[] = [];
-  for (const event of events) {
-    if (
-      phase.length > 0 &&
-      chatEventCompatibilityRole(event.eventType) === "user"
-    ) {
-      phases.push(phase);
-      phase = [];
-    }
-    phase.push(event);
-  }
-  if (phase.length > 0) {
-    phases.push(phase);
-  }
-  return phases;
-}
-
-function lastCompletedWorkEventIndex(
-  events: readonly EnrichedChatEvent[],
-  predicate: (event: EnrichedChatEvent) => boolean,
-): number {
-  for (let index = events.length - 1; index >= 0; index--) {
-    if (predicate(events[index]!)) {
-      return index;
-    }
-  }
-  return -1;
-}
-
-function completedWorkFinalEventIndex(
-  events: readonly EnrichedChatEvent[],
-): number {
-  return lastCompletedWorkEventIndex(events, isRenderableAssistantEvent);
-}
-
-function canFoldCompletedWorkTrailingEvent(event: EnrichedChatEvent): boolean {
-  const role = chatEventCompatibilityRole(event.eventType);
-  return (
-    role === "user" ||
-    (role === "assistant" && !isRenderableAssistantEvent(event))
-  );
-}
-
-interface CompletedWorkPhaseFolding {
-  visibleEvents: readonly EnrichedChatEvent[];
-  fold: CompletedWorkFold | null;
-}
-
-function foldCompletedWorkPhase(
-  runId: string,
-  events: readonly EnrichedChatEvent[],
-): CompletedWorkPhaseFolding {
-  const finalEventIndex = completedWorkFinalEventIndex(events);
-  const finalEvent =
-    finalEventIndex >= 0 ? events[finalEventIndex]! : undefined;
-  const precedingEvents =
-    finalEventIndex > 0 ? events.slice(0, finalEventIndex) : [];
-  const hiddenEvents = precedingEvents.filter((event) => {
-    return (
-      chatEventCompatibilityRole(event.eventType) !== "user" &&
-      !isThinkingOnlyAssistantEvent(event)
-    );
-  });
-  const userEvents = events.filter((event) => {
-    return chatEventCompatibilityRole(event.eventType) === "user";
-  });
-  const trailingEvents =
-    finalEventIndex >= 0 ? events.slice(finalEventIndex + 1) : [];
-  const trailingEventsCanFold = trailingEvents.every((event) => {
-    return canFoldCompletedWorkTrailingEvent(event);
-  });
-  if (
-    finalEvent === undefined ||
-    hiddenEvents.length === 0 ||
-    !trailingEventsCanFold
-  ) {
-    return { visibleEvents: events, fold: null };
-  }
-  return {
-    visibleEvents: [
-      ...userEvents,
-      finalEvent,
-      ...trailingEvents.filter(isRenderableAssistantEvent),
-    ],
-    fold: {
-      key: `${runId}:${finalEvent.id}`,
-      finalEventId: finalEvent.id,
-      hiddenGroups: groupEventsByRole(hiddenEvents),
-      labelGroups: groupEventsByRole(events),
-    },
-  };
-}
-
-function buildCompletedWorkFolding(
-  groups: readonly ChatEventGroup[],
-): CompletedWorkFolding | null {
-  const usageByRunId = usageByRunIdFromGroups(groups);
-  const events = groups.flatMap((group) => {
-    return group.events;
-  });
-  const terminatedRunIds = terminatedRunIdsForCompletedWork(events);
-  const visibleEvents: EnrichedChatEvent[] = [];
-  const folds: CompletedWorkFold[] = [];
-  let hasCompletedWorkPhaseBoundary = false;
-
-  for (let index = 0; index < events.length; ) {
-    const runId = events[index]!.runId;
-    if (runId === undefined) {
-      visibleEvents.push(events[index]!);
-      index++;
-      continue;
-    }
-
-    let endIndex = index + 1;
-    while (endIndex < events.length && events[endIndex]!.runId === runId) {
-      endIndex++;
-    }
-
-    const runEvents = events.slice(index, endIndex);
-    if (!terminatedRunIds.has(runId) || runEvents.some(isCancelledRunEvent)) {
-      visibleEvents.push(...runEvents);
-      index = endIndex;
-      continue;
-    }
-
-    const completedWorkEventGroups = splitCompletedWorkEventsAtUsers(runEvents);
-    if (completedWorkEventGroups.length > 1) {
-      hasCompletedWorkPhaseBoundary = true;
-    }
-    for (const completedWorkEvents of completedWorkEventGroups) {
-      const phaseFolding = foldCompletedWorkPhase(runId, completedWorkEvents);
-      visibleEvents.push(...phaseFolding.visibleEvents);
-      if (phaseFolding.fold !== null) {
-        folds.push(phaseFolding.fold);
-      }
-    }
-
-    index = endIndex;
-  }
-
-  if (folds.length === 0 && !hasCompletedWorkPhaseBoundary) {
-    return null;
-  }
-
-  const foldFinalEventIds = new Set(
-    folds.map((fold) => {
-      return fold.finalEventId;
-    }),
-  );
-  return {
-    visibleGroups: attachUsageToCompletedWorkGroups(
-      groupEventsForCompletedWorkDisplay(visibleEvents, foldFinalEventIds),
-      usageByRunId,
-    ),
-    foldsByFinalEventId: new Map(
-      folds.map((fold) => {
-        return [fold.finalEventId, fold];
-      }),
-    ),
-  };
 }
 
 function parseEventTime(value: string): number | null {
@@ -7553,7 +7181,7 @@ function PagedAssistantEventItem({
 }) {
   const retryRichEventTree = useSet(thread.retryRichEventTree$);
   const pageSignal = useGet(pageSignal$);
-  const error = chatEventError(event);
+  const error = chatEventDisplayError(event);
   if (error) {
     return (
       <ChatAssistantMessageBody
